[Add][Core] Default connection type for providers
What:
A default connection type is set for each provider.
Current types are docker or ssh.
In the future this should allow the  user to override
the default. For example to set ssh for a container.

Why:
This allows the connect.sh script to determine the
type of connection command to invoke, without
requiring that ansible-inventory be present of
the local host.
This means that for container use cases the connect
script does not require ansible-inventory.
In issue #31 there was a difficulty in configuring a CICD
environment with ansible-inventory.

How:
By setting default values in topology-defaults.yml.

Signed-off-by: Begley Brothers Ltd <begleybrothers@gmail.com>